+++ Core/Core.vcxproj.filters
@@ -1175,6 +1175,9 @@
     <ClCompile Include="Debugger\WebSocket\ReplaySubscriber.cpp">
       <Filter>Debugger\WebSocket</Filter>
     </ClCompile>
+    <ClCompile Include="ControlMapper.cpp">
+      <Filter>Core</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ELF\ElfReader.h">
@@ -1891,6 +1894,9 @@
     </ClInclude>
     <ClInclude Include="Debugger\WebSocket\ReplaySubscriber.h">
       <Filter>Debugger\WebSocket</Filter>
+    </ClInclude>
+    <ClInclude Include="ControlMapper.h">
+      <Filter>Core</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
