+++ UWP/CoreUWP/CoreUWP.vcxproj
@@ -382,6 +382,7 @@
     <ClInclude Include="..\..\Core\AVIDump.h" />
     <ClInclude Include="..\..\Core\Compatibility.h" />
     <ClInclude Include="..\..\Core\Config.h" />
+    <ClInclude Include="..\..\Core\ControlMapper.h" />
     <ClInclude Include="..\..\Core\Core.h" />
     <ClInclude Include="..\..\Core\CoreParameter.h" />
     <ClInclude Include="..\..\Core\CoreTiming.h" />
@@ -616,6 +617,7 @@
     <ClCompile Include="..\..\Core\AVIDump.cpp" />
     <ClCompile Include="..\..\Core\Compatibility.cpp" />
     <ClCompile Include="..\..\Core\Config.cpp" />
+    <ClCompile Include="..\..\Core\ControlMapper.cpp" />
     <ClCompile Include="..\..\Core\Core.cpp" />
     <ClCompile Include="..\..\Core\CoreTiming.cpp" />
     <ClCompile Include="..\..\Core\CwCheat.cpp" />
